+++ crates/ryo3-jiff/src/constants.rs
@@ -1,4 +1,5 @@
-use jiff::fmt::temporal::{DateTimeParser, SpanParser};
+use jiff::fmt::temporal::{DateTimeParser, SpanParser, SpanPrinter};
 
 pub(crate) static DATETIME_PARSER: DateTimeParser = DateTimeParser::new();
 pub(crate) static SPAN_PARSER: SpanParser = SpanParser::new();
+pub(crate) static SPAN_PRINTER: SpanPrinter = SpanPrinter::new();