+++ crates/ryo3-jiff/src/ry_datetime.rs
@@ -193,6 +193,17 @@ impl RyDateTime {
         self.0.to_string()
     }
 
+    #[pyo3(name = "to_string")]
+    fn py_to_string(&self) -> String {
+        self.__str__()
+    }
+
+    #[pyo3(
+        warn(
+            message = "obj.string() is deprecated, use `obj.to_string()` or `str(obj)` [remove in 0.0.60]",
+            category = pyo3::exceptions::PyDeprecationWarning
+      )
+    )]
     fn string(&self) -> String {
         self.__str__()
     }