+++ crates/ryo3-jiff/src/ry_time.rs
@@ -124,12 +124,23 @@ impl RyTime {
     // ========================================================================
     // STRING
     // ========================================================================
+    #[pyo3(
+        warn(
+            message = "obj.string() is deprecated, use `obj.to_string()` or `str(obj)` [remove in 0.0.60]",
+            category = pyo3::exceptions::PyDeprecationWarning
+      )
+    )]
     fn string(&self) -> String {
         self.0.to_string()
     }
 
+    #[pyo3(name = "to_string")]
+    fn py_to_string(&self) -> String {
+        self.__str__()
+    }
+
     fn __str__(&self) -> String {
-        self.string()
+        self.0.to_string()
     }
 
     fn __repr__(&self) -> String {