+++ crates/ryo3-jiff/src/ry_zoned.rs
@@ -187,9 +187,19 @@ impl RyZoned {
             CompareOp::Ge => self.0 >= other.0,
         }
     }
-
+    #[pyo3(
+        warn(
+            message = "obj.string() is deprecated, use `obj.to_string()` or `str(obj)` [remove in 0.0.60]",
+            category = pyo3::exceptions::PyDeprecationWarning
+      )
+    )]
     fn string(&self) -> String {
-        self.0.to_string()
+        self.py_to_string()
+    }
+
+    #[pyo3(name = "to_string")]
+    fn py_to_string(&self) -> String {
+        self.__str__()
     }
 
     fn __str__(&self) -> String {