+++ tests/jiff/test_time_span.py
@@ -88,7 +88,7 @@ class TestTimeSpanStrings:
 
     def test_all_ones_repr_full(self) -> None:
         assert (
-            _TIMESPAN_ONES.__repr__()
+            repr(_TIMESPAN_ONES)
             == "TimeSpan(years=1, months=1, weeks=1, days=1, hours=1, minutes=1, seconds=1, milliseconds=1, microseconds=1, nanoseconds=1)"
         )
 
@@ -99,12 +99,12 @@ class TestTimeSpanStrings:
 
     def test_span_str_friendly(self) -> None:
         s = ry.TimeSpan.parse("P2M10DT2H30M")
-        assert s.string(friendly=True) == "2mo 10d 2h 30m"
+        assert s.to_string(friendly=True) == "2mo 10d 2h 30m"
         assert s.friendly() == "2mo 10d 2h 30m"
         assert f"{s:#}" == "2mo 10d 2h 30m"
 
         with pytest.raises(TypeError):
-            assert s.string(True) == "2mo 10d 2h 30m"  # type: ignore[misc] # noqa: FBT003
+            assert s.to_string(True) == "2mo 10d 2h 30m"  # type: ignore[misc] # noqa: FBT003
 
     def test_invalid_format_specifier(self) -> None:
         s = ry.TimeSpan.parse("P2M10DT2H30M")
@@ -113,8 +113,8 @@ class TestTimeSpanStrings:
 
     def test_span_str_alien_or_idk_but_not_human(self) -> None:
         s = ry.TimeSpan.parse("P2M10DT2H30M")
-        assert s.string(friendly=False) == "P2M10DT2H30M"
-        assert s.string() == "P2M10DT2H30M"
+        assert s.to_string(friendly=False) == "P2M10DT2H30M"
+        assert s.to_string() == "P2M10DT2H30M"
 
     def test_repr_kwargs(self) -> None:
         kwarg_keys = (
@@ -133,7 +133,7 @@ class TestTimeSpanStrings:
             s = ry.timespan(**dict.fromkeys(cb, 1))
 
             expected_repr = "TimeSpan(" + ", ".join(f"{k}=1" for k in cb) + ")"
-            assert s.__repr__() == expected_repr
+            assert repr(s) == expected_repr
 
 
 def test_negative_spans() -> None:
@@ -162,19 +162,19 @@ def test_negative_spans() -> None:
     assert_eq!(span.to_string(), "-P5d");
     """
     span = -ry.TimeSpan()._days(5)
-    assert span.string() == "-P5D"
+    assert span.to_string() == "-P5D"
 
     span = ry.TimeSpan()._days(5).negate()
-    assert span.string() == "-P5D"
+    assert span.to_string() == "-P5D"
 
     span = ry.TimeSpan()._days(-5)
-    assert span.string() == "-P5D"
+    assert span.to_string() == "-P5D"
 
     span = -ry.TimeSpan()._days(-5).negate()
-    assert span.string() == "-P5D"
+    assert span.to_string() == "-P5D"
 
     span = ry.TimeSpan()._days(-5)
-    assert span.string() == "-P5D"
+    assert span.to_string() == "-P5D"
 
 
 class TestSpanAdd: